Basic server installation and setup.
My general goal is to make this as hands-off as possible.

On startup, this container pulls down the most recent version of the
dedicated server from terraria's official website. After a bit of extra
setup, it's ready to run the server.

Still need to get it working with a statically provided world file
first. In the future, I want this container to save and load world files
from S3 so we always have a cloudbackup availabe.